Move events() and properties() back under the lifecycle umbrella. They're no longer public API.
