+++ src/index.js
@@ -1,9 +1,7 @@
 import apiCreate from './api/create';
 import apiEmit from './api/emit';
-import apiEvents from './api/events';
 import apiInit from './api/init';
 import apiNoConflict from './api/no-conflict';
-import apiProperties from './api/properties';
 import apiVersion from './api/version';
 import assign from './util/assign';
 import assignSafe from './util/assign-safe';
@@ -105,10 +103,8 @@ function skate (id, userOptions) {
 
 skate.create = apiCreate;
 skate.emit = apiEmit;
-skate.events = apiEvents;
 skate.init = apiInit;
 skate.noConflict = apiNoConflict;
-skate.properties = apiProperties;
 skate.version = apiVersion;
 
 // Global
