+++ src/lifecycle/properties.js
